add catalog sheet component
This commit introduces a new component, `CatalogSheet`. It is a
sheet-based UI element for displaying catalog content. It is designed to
be responsive, using the `useMediaQuery` hook to determine whether to
render a dialog or a sheet based on the screen size. The header
component now uses this new component.